pyproject.toml
[project]
name = "weld-pipeline"
version = "0.1.0"
description = "Industrial-style DE pipeline for welding robot data (synthetic -> staged -> curated -> reports)"
requires-python = ">=3.10"
dependencies = ["pandas", "numpy", "rich", "pydantic", "pyyaml", "streamlit"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
